+++ data_extractor/lxml.py
@@ -2,20 +2,19 @@
 from typing import List, Union
 
 # Third Party Library
-from lxml.etree import _Element as HTMLElement
-from lxml.html import fromstring, tostring
+from lxml.etree import _Element as Element
 
 # Local Folder
 from .abc import AbstractExtractor
 
 
 class CSSExtractor(AbstractExtractor):
-    def extract(self, element: HTMLElement) -> List[HTMLElement]:
+    def extract(self, element: Element) -> List[Element]:
         return element.cssselect(self.expr)
 
 
 class TextCSSExtractor(AbstractExtractor):
-    def extract(self, element: HTMLElement) -> List[str]:
+    def extract(self, element: Element) -> List[str]:
         return [ele.text for ele in CSSExtractor(self.expr).extract(element)]
 
 
@@ -27,7 +26,7 @@ class AttrCSSExtractor(AbstractExtractor):
     def __repr__(self) -> str:
         return f"{self.__class__.__name__}(expr={self.expr!r}, attr={self.attr!r})"
 
-    def extract(self, root: HTMLElement) -> List[str]:
+    def extract(self, root: Element) -> List[str]:
         return [
             ele.get(self.attr)
             for ele in CSSExtractor(self.expr).extract(root)
@@ -36,16 +35,14 @@ class AttrCSSExtractor(AbstractExtractor):
 
 
 class XPathExtractor(AbstractExtractor):
-    def extract(self, element: HTMLElement) -> Union[List["HTMLElement"], List[str]]:
+    def extract(self, element: Element) -> Union[List["Element"], List[str]]:
         return element.xpath(self.expr)
 
 
 __all__ = (
     "AttrCSSExtractor",
     "CSSExtractor",
-    "HTMLElement",
+    "Element",
     "TextCSSExtractor",
     "XPathExtractor",
-    "fromstring",
-    "tostring",
 )
